refactor: simplify KustomizeRenderer
