--- src/components/organisms/ExplorerPane/DryRunsPane/KustomizeRenderer/index.ts
@@ -1,1 +1,0 @@
-export {default} from './KustomizeRenderer';
